LPD-82692 Move collection-display and mapping tests into shared specs
